Light-mode: CLI options
