+++ cabal.project
@@ -38,7 +38,7 @@
 --
 --------------------------------------------------------------------------------
 
-index-state: 2022-01-01T00:00:00Z
+index-state: 2022-02-22T20:47:03Z
 with-compiler: ghc-8.10.7
 
 packages:
