+++ flake.nix
@@ -49,8 +49,12 @@
       flake = false;
     };
     customConfig.url = "github:input-output-hk/empty-flake";
-    emanote.url = "github:srid/emanote";
-    emanote.inputs.nixpkgs.url = github:NixOS/nixpkgs/d77bbfcbb650d9c219ca3286e1efb707b922d7c2;
+    emanote = {
+      url = "github:srid/emanote";
+    };
+    ema = {
+      url = "github:srid/ema";
+    };
   };
 
   outputs = { self, nixpkgs, hostNixpkgs, flake-utils, haskellNix, iohkNix, customConfig, emanote, ... }:
